split download primitives into bitnet-download-core microcrate
