+++ crates/bitnet-download-core/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "bitnet-download-core"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+documentation.workspace = true
+description = "Core SRP download primitives for offline mode, byte-range parsing, and atomic metadata writes"
+
+[dependencies]
+thiserror.workspace = true
+
+[lints]
+workspace = true
